+++ Server/include/LKZ/Protocol/Message/Approach/DisconnectClientMessage.h
@@ -5,11 +5,11 @@
 
 struct DisconnectClientMessage : public Message
 {
-    static constexpr int ID = 3;
+    static constexpr uint8_t ID = 3;
 
     DisconnectClientMessage();
 
-    byte getId() const override;
+    uint8_t getId() const override;
 
     std::vector<uint8_t>& serialize(Serializer& serializer) const override;
     void deserialize(Deserializer& deserializer) override;
