+++ Server/include/LKZ/Protocol/Message/Matchmaking/StartGameMessage.h
@@ -5,13 +5,13 @@
 
 struct StartGameMessage : public Message
 {
-    static constexpr int ID = 5;
+    static constexpr uint8_t ID = 5;
 
     StartGameMessage();
 
-    byte mapId;
+    uint8_t mapId;
 
-    byte getId() const override;
+    uint8_t getId() const override;
 
     std::vector<uint8_t>& serialize(Serializer& serializer) const override;
     void deserialize(Deserializer& deserializer) override;
